ajout moteur RPN et conversion au type double

+++ src/myRPNcalc/Commande.java
@@ -3,5 +3,5 @@ package myRPNcalc;
 @FunctionalInterface
 public interface Commande {
 
-	public int apply();
+	public Double apply();
 }
